# internal/httpapi/openapi.yaml
openapi: 3.1.0
info:
  title: managed-agent-go
  version: 0.1.0
  description: >-
    Clean-room implementation of the supported Claude Managed Agents API core,
    conditional Files and Memory APIs, and File- and Memory-backed Session
    Resources. This project is Apache-2.0 licensed and is not an Anthropic
    product.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: /
    description: Current managed-agent-go server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
  - {}
...
paths:
...
  /v1/sessions:
    post:
      operationId: createSession
      tags: [Sessions]
      summary: Create a Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionCreateRequest'
      responses:
        '200':
          $ref: '#/components/responses/SessionResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          $ref: '#/components/responses/RequestTooLarge'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
    get:
      operationId: listSessions
      tags: [Sessions]
      summary: List Sessions
      parameters:
        - $ref: '#/components/parameters/SessionListLimit'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Order'
        - $ref: '#/components/parameters/AgentIDFilter'
        - $ref: '#/components/parameters/AgentVersionFilter'
        - $ref: '#/components/parameters/Statuses'
        - $ref: '#/components/parameters/IncludeArchived'
        - $ref: '#/components/parameters/CreatedAtGt'
        - $ref: '#/components/parameters/CreatedAtGte'
        - $ref: '#/components/parameters/CreatedAtLt'
        - $ref: '#/components/parameters/CreatedAtLte'
        - $ref: '#/components/parameters/DeploymentIDFilter'
        - $ref: '#/components/parameters/MemoryStoreIDFilter'
      responses:
        '200':
          $ref: '#/components/responses/SessionListResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
  /v1/sessions/{session_id}:
    parameters:
      - $ref: '#/components/parameters/SessionID'
    get:
      operationId: getSession
      tags: [Sessions]
      summary: Get a Session
      responses:
        '200':
          $ref: '#/components/responses/SessionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      operationId: updateSession
      tags: [Sessions]
      summary: Update Session metadata, title, or the next-turn tool configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionUpdateRequest'
      responses:
        '200':
          $ref: '#/components/responses/SessionResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
    delete:
      operationId: deleteSession
      tags: [Sessions]
      summary: Delete a non-running Session and its event history
      responses:
        '200':
          $ref: '#/components/responses/SessionDeletedResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /v1/sessions/{session_id}/resources:
    parameters:
      - $ref: '#/components/parameters/SessionID'
    post:
      operationId: addSessionResource
      tags: [Session Resources]
      summary: Attach a File copy to a Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileSessionResourceInput'
      responses:
        '200':
          $ref: '#/components/responses/SessionResourceResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/RequestTooLarge'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
    get:
      operationId: listSessionResources
      tags: [Session Resources]
      summary: List Session Resources
      parameters:
        - $ref: '#/components/parameters/SessionResourceListLimit'
        - $ref: '#/components/parameters/Page'
      responses:
        '200':
          $ref: '#/components/responses/SessionResourceListResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
  /v1/sessions/{session_id}/resources/{resource_id}:
    parameters:
      - $ref: '#/components/parameters/SessionID'
      - $ref: '#/components/parameters/SessionResourceID'
    get:
      operationId: getSessionResource
      tags: [Session Resources]
      summary: Get a Session Resource
      responses:
        '200':
          $ref: '#/components/responses/SessionResourceResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
    post:
      operationId: updateSessionResource
      tags: [Session Resources]
      summary: Update a Session Resource
      description: File resources reject this GitHub-only token rotation operation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionResourceUpdateRequest'
      responses:
        '200':
          $ref: '#/components/responses/SessionResourceResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
    delete:
      operationId: deleteSessionResource
      tags: [Session Resources]
      summary: Detach a File Resource from a Session
      responses:
        '200':
          $ref: '#/components/responses/SessionResourceDeletedResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnsupportedRequest'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
    BearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    SessionID:
      name: session_id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
...
    SessionListLimit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
...
    Page:
      name: page
      in: query
      description: Opaque cursor returned by a previous list response.
      schema:
        type: string
        minLength: 1
    IncludeArchived:
      name: include_archived
      in: query
      schema:
        type: boolean
        default: false
    Order:
      name: order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: desc
    AgentIDFilter:
      name: agent_id
      in: query
      schema:
        type: string
    AgentVersionFilter:
      name: agent_version
      in: query
      description: Requires agent_id.
      schema:
        type: integer
        minimum: 1
    Statuses:
      name: statuses[]
      in: query
      style: form
      explode: true
      schema:
        type: array
        items:
          $ref: '#/components/schemas/SessionStatus'
    DeploymentIDFilter:
      name: deployment_id
      in: query
      description: Accepted for compatibility; no current Session matches a deployment.
      schema:
        type: string
    MemoryStoreIDFilter:
      name: memory_store_id
      in: query
      description: Return Sessions attached to this Memory Store.
      schema:
        type: string
        pattern: '^memstore_'
    SessionResourceID:
      name: resource_id
      in: path
      required: true
      schema:
        type: string
        pattern: '^sesrsc_'
    SessionResourceListLimit:
      name: limit
      in: query
      description: Maximum resources per page. Omit to return all Session Resources.
      schema:
        type: integer
        minimum: 1
        maximum: 1000
    CreatedAtGt:
      name: created_at[gt]
      in: query
      schema:
        type: string
        format: date-time
    CreatedAtGte:
      name: created_at[gte]
      in: query
      schema:
        type: string
        format: date-time
    CreatedAtLt:
      name: created_at[lt]
      in: query
      schema:
        type: string
        format: date-time
    CreatedAtLte:
      name: created_at[lte]
      in: query
      schema:
        type: string
        format: date-time
  responses:
...
    SessionResponse:
      description: Session resource
      headers:
        request-id:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Session'
    SessionListResponse:
      description: Session page
      headers:
        request-id:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SessionList'
    SessionResourceResponse:
      description: File- or Memory-backed Session Resource
      headers:
        request-id:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SessionResource'
    SessionResourceListResponse:
      description: Session Resource page
      headers:
        request-id:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SessionResourceList'
    SessionResourceDeletedResponse:
      description: Session Resource deletion confirmation
      headers:
        request-id:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SessionResourceDeleted'
    SessionDeletedResponse:
      description: Session deleted
      headers:
        request-id:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SessionDeleted'
...
    InvalidRequest:
      description: Invalid request
      headers:
        request-id:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      headers:
        request-id:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: State or optimistic-concurrency conflict
      headers:
        request-id:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnsupportedRequest:
      description: Well-formed capability that this implementation cannot enforce
      headers:
        request-id:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RequestTooLarge:
      description: Request exceeds the endpoint size limit
      headers:
        request-id:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  headers:
    RequestID:
      description: Identifier for this API request.
      schema:
        type: string
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [type, error, request_id]
      properties:
        type:
          type: string
          const: error
        error:
          type: object
          additionalProperties: false
          required: [type, message]
          properties:
            type:
              type: string
              enum:
                - invalid_request_error
                - authentication_error
                - not_found_error
                - conflict_error
                - memory_precondition_failed_error
                - request_too_large
                - api_error
            message:
              type: string
        request_id:
          type: string
...
    Metadata:
      type: object
      maxProperties: 16
      propertyNames:
        maxLength: 64
      additionalProperties:
        type: string
        maxLength: 512
    MetadataPatch:
      type: object
      propertyNames:
        maxLength: 64
      additionalProperties:
        oneOf:
          - type: string
            maxLength: 512
          - type: 'null'
    NullableTimestamp:
      type: [string, 'null']
      format: date-time
    Cursor:
      type: [string, 'null']
...
    AgentTool:
      oneOf:
        - $ref: '#/components/schemas/BuiltinToolset'
        - $ref: '#/components/schemas/MCPToolset'
        - $ref: '#/components/schemas/CustomTool'
      discriminator:
        propertyName: type
    MCPServer:
      type: object
      additionalProperties: false
      required: [type, name, url]
      properties:
        type:
          type: string
          const: url
        name:
          type: string
          minLength: 1
          maxLength: 255
        url:
          type: string
          format: uri
          maxLength: 2048
...
    AgentSnapshot:
      type: object
      additionalProperties: false
      required:
        - id
        - type
        - version
        - name
        - model
        - system
        - description
        - tools
        - mcp_servers
        - skills
        - multiagent
      properties:
        id:
          type: string
        type:
          type: string
          const: agent
        version:
          type: integer
          minimum: 1
        name:
          type: string
        model:
          $ref: '#/components/schemas/Model'
        system:
          type: string
        description:
          type: string
        tools:
          type: array
          items:
            $ref: '#/components/schemas/AgentTool'
        mcp_servers:
          type: array
          items:
            $ref: '#/components/schemas/MCPServer'
        skills:
          type: array
          items:
            $ref: '#/components/schemas/SkillReferenceResponse'
        multiagent:
          $ref: '#/components/schemas/Multiagent'
...
    AgentReference:
      type: object
      additionalProperties: false
      required: [type, id]
      properties:
        type:
          type: string
          const: agent
        id:
          type: string
          minLength: 1
        version:
          type: integer
          minimum: 1
    AgentWithOverrides:
      type: object
      additionalProperties: false
      required: [type, id]
      properties:
        type:
          type: string
          const: agent_with_overrides
        id:
          type: string
          minLength: 1
        version:
          type: integer
          minimum: 1
        model:
          $ref: '#/components/schemas/ModelInput'
        system:
          type: [string, 'null']
        tools:
          oneOf:
            - type: array
              maxItems: 128
              items:
                $ref: '#/components/schemas/AgentTool'
            - type: 'null'
        mcp_servers:
          oneOf:
            - type: array
              maxItems: 20
              items:
                $ref: '#/components/schemas/MCPServer'
            - type: 'null'
        skills:
          oneOf:
            - type: array
              maxItems: 500
              items:
                $ref: '#/components/schemas/SkillReferenceInput'
            - type: 'null'
    SessionAgentInput:
      oneOf:
        - type: string
          minLength: 1
        - $ref: '#/components/schemas/AgentReference'
        - $ref: '#/components/schemas/AgentWithOverrides'
...
    MessageContentInput:
      oneOf:
        - $ref: '#/components/schemas/TextBlockInput'
        - $ref: '#/components/schemas/ImageBlockInput'
        - $ref: '#/components/schemas/DocumentBlockInput'
      discriminator:
        propertyName: type
    UserMessageInitialEvent:
      type: object
      additionalProperties: false
      required: [type, content]
      properties:
        type:
          type: string
          const: user.message
        content:
          type: array
          minItems: 1
          maxItems: 1000
          items:
            $ref: '#/components/schemas/MessageContentInput'
...
    DefineOutcomeInitialEvent:
      type: object
      additionalProperties: false
      required: [type, description, rubric]
      properties:
        type:
          type: string
          const: user.define_outcome
        description:
          type: string
          minLength: 1
        rubric:
          $ref: '#/components/schemas/OutcomeRubric'
        max_iterations:
          type: integer
          minimum: 1
          maximum: 20
    InitialEvent:
      oneOf:
        - $ref: '#/components/schemas/UserMessageInitialEvent'
        - $ref: '#/components/schemas/DefineOutcomeInitialEvent'
      discriminator:
        propertyName: type
    SessionCreateRequest:
      type: object
      additionalProperties: false
      required: [agent, environment_id]
      properties:
        agent:
          $ref: '#/components/schemas/SessionAgentInput'
        environment_id:
          type: string
          minLength: 1
        title:
          type: string
        metadata:
          $ref: '#/components/schemas/Metadata'
        initial_events:
          type: array
          maxItems: 50
          items:
            $ref: '#/components/schemas/InitialEvent'
        resources:
          type: array
          maxItems: 500
          description: At most eight entries may be Memory Stores; Memory Stores attach only at creation.
          items:
            $ref: '#/components/schemas/SessionResourceInput'
        vault_ids:
          type: array
          maxItems: 0
          items:
            type: string
    SessionAgentUpdate:
      type: object
      additionalProperties: false
      properties:
        tools:
          oneOf:
            - type: array
              maxItems: 128
              items:
                $ref: '#/components/schemas/AgentTool'
            - type: 'null'
        mcp_servers:
          oneOf:
            - type: array
              maxItems: 20
              items:
                $ref: '#/components/schemas/MCPServer'
            - type: 'null'
    SessionUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
        metadata:
          $ref: '#/components/schemas/MetadataPatch'
        agent:
          $ref: '#/components/schemas/SessionAgentUpdate'
        vault_ids:
          description: Reserved and rejected with 422 when present.
          type: array
          items:
            type: string
    SessionStatus:
      type: string
      enum: [idle, running, rescheduling, terminated]
    FileSessionResourceInput:
      type: object
      additionalProperties: false
      required: [type, file_id]
      properties:
        type:
          type: string
          const: file
        file_id:
          type: string
          minLength: 1
        mount_path:
          type: string
          minLength: 1
          description: Absolute path normalized beneath /mnt/session/uploads.
    MemoryStoreSessionResourceInput:
      type: object
      additionalProperties: false
      required: [type, memory_store_id]
      properties:
        type:
          type: string
          const: memory_store
        memory_store_id:
          type: string
          pattern: '^memstore_'
        access:
          type: string
          enum: [read_write, read_only]
          default: read_write
        instructions:
          type: string
          maxLength: 4096
    SessionResourceInput:
      oneOf:
        - $ref: '#/components/schemas/FileSessionResourceInput'
        - $ref: '#/components/schemas/MemoryStoreSessionResourceInput'
      discriminator:
        propertyName: type
...
    SessionResource:
      oneOf:
        - $ref: '#/components/schemas/FileSessionResource'
        - $ref: '#/components/schemas/MemoryStoreSessionResource'
      discriminator:
        propertyName: type
    SessionResourceUpdateRequest:
      type: object
      additionalProperties: false
      required: [authorization_token]
      properties:
        authorization_token:
          type: string
          minLength: 1
          description: GitHub-only field; File resources return 422.
    SessionResourceDeleted:
      type: object
      additionalProperties: false
      required: [id, type]
      properties:
        id:
          type: string
          pattern: '^sesrsc_'
        type:
          type: string
          const: session_resource_deleted
    SessionResourceList:
      type: object
      additionalProperties: false
      required: [data, has_more, next_page]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/SessionResource'
        has_more:
          type: boolean
        next_page:
          $ref: '#/components/schemas/Cursor'
    SessionStats:
      type: object
      additionalProperties: false
      required: [active_seconds, duration_seconds]
      properties:
        active_seconds:
          type: number
          minimum: 0
        duration_seconds:
          type: number
          minimum: 0
    SessionUsage:
      type: object
      additionalProperties: false
      required: [cache_creation, cache_read_input_tokens, input_tokens, output_tokens]
      properties:
        cache_creation:
          type: object
          additionalProperties: false
          required: [ephemeral_1h_input_tokens, ephemeral_5m_input_tokens]
          properties:
            ephemeral_1h_input_tokens:
              type: integer
              minimum: 0
            ephemeral_5m_input_tokens:
              type: integer
              minimum: 0
        cache_read_input_tokens:
          type: integer
          minimum: 0
        input_tokens:
          type: integer
          minimum: 0
        output_tokens:
          type: integer
          minimum: 0
    OutcomeEvaluation:
      type: object
      additionalProperties: false
      required: [type, outcome_id, description, result, explanation, iteration, completed_at]
      properties:
        type:
          type: string
          const: outcome_evaluation
        outcome_id:
          type: string
        description:
          type: string
        result:
          type: string
        explanation:
          type: [string, 'null']
        iteration:
          type: integer
          minimum: 0
        completed_at:
          $ref: '#/components/schemas/NullableTimestamp'
    Session:
      type: object
      additionalProperties: false
      required:
        - id
        - type
        - status
        - agent
        - environment_id
        - title
        - metadata
        - outcome_evaluations
        - resources
        - stats
        - usage
        - vault_ids
        - deployment_id
        - created_at
        - updated_at
        - archived_at
      properties:
        id:
          type: string
        type:
          type: string
          const: session
        status:
          $ref: '#/components/schemas/SessionStatus'
        agent:
          $ref: '#/components/schemas/AgentSnapshot'
        environment_id:
          type: string
        title:
          type: string
        metadata:
          $ref: '#/components/schemas/Metadata'
        outcome_evaluations:
          type: array
          items:
            $ref: '#/components/schemas/OutcomeEvaluation'
        resources:
          type: array
          maxItems: 500
          items:
            $ref: '#/components/schemas/SessionResource'
        stats:
          $ref: '#/components/schemas/SessionStats'
        usage:
          $ref: '#/components/schemas/SessionUsage'
        vault_ids:
          type: array
          maxItems: 0
          items:
            type: string
        deployment_id:
          type: 'null'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        archived_at:
          $ref: '#/components/schemas/NullableTimestamp'
    SessionList:
      type: object
      additionalProperties: false
      required: [data, next_page, prev_page]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Session'
        next_page:
          $ref: '#/components/schemas/Cursor'
        prev_page:
          $ref: '#/components/schemas/Cursor'
    SessionDeleted:
      type: object
      additionalProperties: false
      required: [id, type]
      properties:
        id:
          type: string
        type:
          type: string
          const: session_deleted
...